feat: Add 'TupleOf' utility type.
