+++ packages/utility/src/types/utility.ts
@@ -1,3 +1,4 @@
+export { Spread } from './utility/spread';
 export type EmptyArray = never[];
 export type EmptyObject = Record<string, never>;
 export type HttpMethod =
@@ -82,33 +83,14 @@ export type OmitByValue<Container, ValueTypeFilter> = Omit<
   ExtractKeysByValue<Container, ValueTypeFilter>
 >;
 
-// SPREAD (Thanks to https://stackoverflow.com/questions/49682569/typescript-merge-object-types)
-
-type OptionalPropertyNames<T> = {
-  [K in keyof T]-?: {} extends { [P in K]: T[K] } ? K : never;
-}[keyof T];
-
-type SpreadProperties<L, R, K extends keyof L & keyof R> = {
-  [P in K]: L[P] | Exclude<R[P], undefined>;
-};
-
-type Id<T> = T extends infer U ? { [K in keyof U]: U[K] } : never;
-
-type SpreadTwo<L, R> = Id<
-  Pick<L, Exclude<keyof L, never | keyof R>> &
-    Pick<R, Exclude<keyof R, OptionalPropertyNames<R>>> &
-    Pick<R, Exclude<OptionalPropertyNames<R>, keyof L>> &
-    SpreadProperties<L, R, OptionalPropertyNames<R> & keyof L>
->;
-
-export type Spread<A extends readonly [...any]> = A extends [
-  infer L,
-  ...infer R
-]
-  ? SpreadTwo<L, Spread<R>>
-  : unknown;
-
-// END SPREAD
+export type TupleOf<T, N extends number> = N extends N
+  ? number extends N
+    ? T[]
+    : _TupleOf<T, N, []>
+  : never;
+type _TupleOf<T, N extends number, R extends unknown[]> = R['length'] extends N
+  ? R
+  : _TupleOf<T, N, [T, ...R]>;
 
 /**
  * Primitive type
